Fix : Stat, Reduce Quantity, Add Transactions

+++ app/src/main/java/com/fajarraya/app/components/cards/StatsCard.kt
@@ -23,7 +23,7 @@ import com.fajarraya.app.constants.WidgetConstants
 fun StatsCard(
     modifier : Modifier = Modifier,
     icons : Int,
-    totalStats : Int,
+    totalStats : Long,
     statsText : String,
     cardColor : Color
