+++ app/src/main/java/com/fajarraya/app/models/Stats.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.graphics.Color
 
 data class Stats(
     val statsText : String,
-    val total : Int,
+    val total : Long,
     val icons : Int,
     val color : Color
 )
